% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.packages.from.anywhere.R
\name{install.packages.from.anywhere}
\alias{install.packages.from.anywhere}
\title{Install R Packages from Anywhere (CRAN, Bioconductor, GitHub)}
\usage{
install.packages.from.anywhere(
  packages,
  github_map = NULL,
  update = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{packages}{Character vector of package names}

\item{github_map}{Named vector mapping package names to GitHub repos ("owner/repo")}

\item{update}{Logical, whether to update dependencies in Bioconductor}

\item{verbose}{Logical, whether to print progress messages}
}
\value{
Character vector of packages not found
}
\description{
Attempts to install packages from CRAN, then Bioconductor, and finally GitHub.
If a package cannot be found, it is returned in a vector.
}
